CHANNELMGR-1946 Separate channel state in ExtJs and AngularJs
ExtJs now passes a channel ID to the AngularJs app instead of a complete
channel JSON blob. The app then fetches the channel information itself.
This removes the need for all kinds of events between ExtJs and the app.
The initial flow is now as follows:

1. The "branched channel id" (e.g. the channel's configuration name
   including the branchId like "mychannel-branchId-preview") is already
   calculated in Java. This avoids the need to (re)calculate the
   'channelBaseId' in the ProjectService, which avoids the need to fetch
   all channels whenever a channel is initialized.

2. ExtJs passes the channelId to the app.

3. The hippoCm controller initializes the ChannelService and (if needed)
   the URL in the iframe afterwards, or only the iframe directly when
   only the path changed.

4. The app fetches the channel info and (if not passed) the active
   project. The backend is then initialized and eventually the current
   channel is set in the ChannelService.

5. The app sends a 'set-breadcrumb' to ExtJs to update the channel name
   in the breadcrumb.

The app no longer sends a load-channel, reload-channel or switch-channel
event to ExtJs. The only channel-related data needed in ExtJs is the
mountId and contextPath of the current channel when opening the
component properties dialog, which are passed when the dialog is opened.

Still TODO: fix the unit tests of ChannelService and ProjectService.