HSTTWO-4383 Move the RequestContextProvider to shared lib
The RequestContextProvider is a major fundamental utility used all around
in the HST and by end projects to get hold of the HstRequestContext.
However at this moment, the RequestContextProvider is part of hst-commons,
and hst-commons jar is deployed in every webapp. This means that if a
request for the /site webapp is processed, and then in a platform
controlled object (for example HstLinkCreator) the code

RequestContextProvider.get()

then the classloader of RequestContextProvider is the classloader from
the cms webapp but the HstRequestContext is set on the
RequestContextProvider of the site webapp.
To circumvent this problem, the RequestContextProvider can be moved to
the hst-api that is deployed in the shared lib.
However, then we have the next challenge: It is not possible for invoke
the package private static method

org.hippoecm.hst.container.RequestContextProvider#set

from for example HstDelegateeFilterBean although HstDelegateeFilterBean
lives in the same package. The reason for this is that
HstDelegateeFilterBean is loaded with a different classloader than
RequestContextProvider causing the #set to result in an IllegalAccessError.

I don't want to make the

org.hippoecm.hst.container.RequestContextProvider#set

public since then everyone sees this setter when using the
RequestContextProvider.
A way to circumvent this is to add a protected static innerclass in
RequestContextProvider that if you extend if (only possible from same
package), you can invoke set(HstRequestContext). Thus to set the request
context, you can only do so from the same package using:

    private final RequestContextProvider.ModifiableRequestContextProvider
     modifiableRequestContextProvider =
            new RequestContextProvider.ModifiableRequestContextProvider() {};
            modifiableRequestContextProvider.set(requestContext);

For example removing the 'protected' modifier from the static innerclass
 in the shared lib already results in an IllegalAccessError, just like
 pretty much any other change. Either way, this solution provides
 exactly what we need.