CHANNELMGR-1946 Remove channelId from hippo-cm.channel state
It's not used; code that needs the current channel fetches it from the
ChannelService instead of the uiRouter state. Simplifies the state even
further: the hippo-cm-channel state means "show the channel editor".
The currently active channel can change in various ways (e.g. from the
outside by selecting one in the overview, view-document-in-channel or
via Projects; or by following a cross-channel link). All this only
updates the various services, and the views re-render because they bind
to data in those services.

A consequence of this is that following a cross-channel link while the
right side-panel is open keeps the right side-panel open. This is also
an improvement, since things done/shown in the right side panel are
independent of the channel currently shown.