CHANNELMGR-1946 Reload app state when perspective is activated
Reset the ExtJs channel state directly after sending the load-channel
event. Subsequent loadChannel invocations (because the perspective is
activated) will send a reload-channel event that refreshes the current
app state (which may be a different channel and/or project at that time)

Because the whole channel state is refreshed whenever the perspective is
reactivated, the Projects-specific event bus is not needed anymore. The
app simply refreshes its view, so changed project details or state are
directly shown too.