REPO-2072 Only skip logging for idempotant invocations
We should not skip logging for workflow invocations that modify the
state in the repository. For idempotent calls like #listBranches and
#getBranch which do not modify state it is better to skip logging.
However for something like #reintegrateBranch or #publishBranch it
really is not ok to skip logging: AFAIK the logging results in
hippo:log entries and are also used for audit logging. Just because we
do not want to see these workflow steps in the activity stream in the cms
doesn't mean we should suppress logging. It is like
ObtainEditableInstance : I'd think this just gets logged but it doesn't
show up in activity stream either.