CHANNELMGR-1946 Initialize current project with 'core'
The community edition never calls ProjectService.load(), so the active
project (this.project) was never set. Yet a lot of checks are done on
this.project, so in the community edition, so this.project should be
this.core by default.

Also initialize allProjects with an empty array ("no projects") so
methods that do this.allProjects.find() won't break.