CHANNELMGR-1972 Reuse default picker properties
I refactored the picker properties even further, so it becomes more
clear what the differences are and which still need to be added, like
"base.path" for image pickers (I was wrong about it missing for image
pickers) and enable.upload for the RichTextImagePicker.

I added the 'RichText' prefix to the rich-text-pickers and changed the
removePrefix functionality into simply prefix: previously the properties
where configured with a prefix that was removed again when storing it in
JSON. This prohibits use to reuse the properties, so I inverted the
behavior; if a prefix is set, it will be added to the property name when
looking up a JCR property.