CHANNELMGR-1979 Fix the md-select model on open due to scope bug
The select showed 2 values, the previously selected project and core,
even though the previously selected project was not even in the project
array.
After several ways of trying this workaround of setting the projects
array on open solved the problem. Its not ideal as the transition seems
less smooth but it will have to do for now.