CMS-11297 Replace observer by shared model
Introduce a BranchIdModel that wraps the ReferenceModel containing
the branchId in order to have all logic in one place.

Remove the whole observer pattern, because the components detect
changes in the BranchIdModel.

Replace setting the initial version of the branch id in each component
by setting the initial version in the BrowserService, because that
is the point where a the BranchId of the node to be viewed / edited
can be determined the earliest.

Destroy the branchId model when the editor is closed instead of
keeping the model in memory, so that the editor always start with
the branchId of the unpublished node. This limits the complexity.

Detach unpublished, published and draft node models when possible
to prevent undetached models.