CHANNELMGR-1994 Use correct context path per channel
With multiple site web apps, the context path per channel differs. REST
calls should hit the webapp that contains the current channel, and hence
use the contextPath property of a channel. Yet the channel overview only
passes the channelId to the Angular app. To fetch a channel's properties
we need the channel's context path in the REST URL, but that context
path is part of the channel's properties: a catch 22.

Fixed by also passing a channel's context path to the Angular app in the
'load-channel' event. Fetching a channel is now done with the context
path explicitly provided to HstService.getChannel(). Cross-channel links
will also pass the context path of the new channel, and so does the page
copy action.