CHANNELMGR-1994 Reload current page when possible
Reload keeps the scroll position, so reload the same page whenever
possible. It's not possible when either the renderPath differs
(so we have to show a different page) or the renderPath is the same
but the contextPath differs (so we show the exact same renderPath but in
a different channel from a different webapp).

Also added a unit test for the dev-mode specific storage of the current
path in sessionStorage.