CMS-11309 Move initialization of branchIdModel up
An implementation of an abstact method in  subclass
of AbstractPreviewWorkflowPlugin needs access to the branchIdModel.

The constructor of the AbstractPReviewWorkflowPlugin calls the
implementation. At that time the constructor of the subclass
hasn't yet finished and so the branchIdModel has not been initialized.

That why the initialized has been moved up.