CHANNELMGR-2004 Fix opening a channel branch
The context path of the channel to open has to be fetched from the
channel store of the overview. However, that store only contains the
master channels (the default one or the -preview version if it exists),
but not the branches. We therefore pass the unmodified channelId to the
frontend, so it can use it to find the channel in the store and its
context path. Optimization: when there's only one context path (e.g. 99%
of the cases) just use that one.