CHANNELMGR-2004 Retain sub-state when loading a channel
When a channel has already been loaded, a user may already be editing a
document, creating a document etc. These are all sub-states of
"hippo-cm.channel". This sub-state can be kept, since it's independent
of the channel that's shown. However, when the hippo-cm.channel state is
loaded explicitly, the right side-panel closes. Avoid that by keeping
the sub-state.