CHANNELMGR-1999 Confirm closing dirty content on project switch
The ProjectService got 'before' and 'after' change listeners that are
called whenever a different project is selected. Both return promises
that should all resolve. Both the EditContent and CreateContent service
use a 'before' change listener to show a confirmation dialog for dirty
content. 'Save' saves changes, 'Discard' discards them, and 'Cancel'
rejects the promise chain so the selected project does not change.

Projects can also be switched via the load-channel event. This triggers
ProjectService.load(), which now also calls the before and after
listeners to trigger confirmation dialogs for dirty content. A 'Cancel'
will reject the promise chain in ChannelService.initializeChannel() so
the new channel won't be loaded. In that case no error is logged anymore
since it's a valid UI path. In that case the session for an existing
channel should be restored, so subsequent REST calls will hit the right
webapp.