CMS-11311 Use BranchHandle to get variant or frozen node
Use BranchHandle to determine if a variant should be read from the
handle or from from version history.

The workflow and the branchId can be null ( if called from the unit
tests ( that have not been touched ) ). In that case the variant is
always returned. That should never occur outside the testing scope.

Added branchId and Workflow to all unit tests would introduce a lot
of mocking without adding a lot of extra coverage. BranchHandle itself
already has unit tests.