CHANNELMGR-2009 Remove "Core" concept
Remove "Core" object from service, controler and component, because
master is now part of the projects that are returned by the
backend calls.

Remove translation key, because master is now localized by the
backend.

The "master" id is still needed, because when the channel is loaded
it needs a default project if branchId has not been supplied and no
project has already been selected. Assume that there is always a
"Core" branch send by the backend.