CHANNELMGR-2009 Remove "Core" concept (last parts)
Now that the back-end returns the master project as well when calling
associated-with-channel we no longer have to get all projects. So we
can remove the _getAllProjects() code and usages. The returned projects
are not sorted, so we need to do that client-side and make sure that
the master project is the first one.