REPO-2085 Make hints aware of branchId
Make the DocumentWorkflow hints aware of the current branchId, so
the hints can be based on information from the variants under the handle
and the frozen nodes.

Make BranchHandle deal with the situation that there is no draft.

Create separate BranchHints class and relevant test class to reduce the
impact on existing integration tests.

Add assertJ to simplify assertions based on maps like the workflow
hints.