CMS-11347 Make PublicationWorkflowPlugin aware of branches
In order to make the PublicationWorkflowPlugin aware of branches, the
hints had to be a function of the "current branch id" ( The branch
that is currently viewed in the document editor without the need to
be copied to the unpublised variant under the handle ).

AbstractDocumentWorkflow#getHints method is used for all plugins
dealing with DocumentWorkflow.

Make the hints a map of hints to be able to cache hints for various
branches. The hints are a function of the branch id. See REPO-2085.

If the user selects a publish action while viewing a frozen node (
that can only happen in case the "current branch id" is master ), then
a checkout is necessary, because publication is only allowed for the
master version. The user will not notice the checkout action, because
the frozen nodes in that case have the same content as the nodes
under the handle.

For now the (de)publish and scheduled(de)Publication use the same hints.
This has to be changed in the near future. (See REPO-2029)