HSTTWO-4398 Account for missing versionHistory property
The HST should try to find the best matching node for rendering a
document. If the versionHistory mixin is present, but the property is
not, then we can't lookup the node in version history. In that case
the best matching node representing the document contenst is the node
itself.

This situation occurs after upgrading from a pre-12.5 version to 12.5.
The mixin migrator (from WPM_847) cannot set the versionHistory
property on branched documents, so the hippo:versionInfo mixin is set
and the hippo:branches property too, but the hippo:versionHistory
property is absent. Other possible situation is that the property was
removed (by accident?) via the console by an admin user.