CMS-11352 Remove TransparentFragment
The logic of TransparentFragment does not work anymore in Wicket 7. It
relied on the logic inside AbstractMarkupSourcingStrategy to resolve
markup in transparent resolvers, which changed completely in Wicket 7.
As a result, the child inside a TransparentFragment could not be found
anymore in an Ajax response.

TransparentFragment was already marked as deprecated, so the original
author probably suspected it would break somewhere in the future...

The only usage of TransparentFragment was in NodeFieldPlugin and
PropertyFieldPlugin, which use fragments to add the markup for 'view'
and 'edit' mode. That markup has now been refactored into separate
components:

 - FieldContainer shares the common markup for all field containers
   (both for view and edit mode)
 - EditableFieldContainer shares the common markup for all field
   containers in edit mode
 - EditableNodeFieldContainer and EditablePropertyFieldContainer have
   the logic for editing the order of node and property fields.