REPO-2085 Make hints use the variants from the Branch handle
Combine the logic of the documentworkflow.scxml with the logic to
select a node from version history or from under the handle for deter-
mining the right hints. The resulting hints are the same as the hints
if the possible versioned nodes were copied under the handle.

Introduce a separate workflow executor because most of the actions
do not operate on versioned nodes.