REPO-2089 Just read the property instead of checking mixin presence
The backing node can be a frozen node. Frozen nodes don't have mixins
except for the mix:referenceable. So checking for mixin presence does
not work. To fix the issue we can just read the property and return
MASTER if it's absent.

To verify my fix I modified a test to use the document variant method.
I also added the method getBranchName, because we'll need it in CMS
code.