CHANNELMGR-1973 Re-evaluate hints for non-master documents
Javadoc in DocumentServiceImpl states that the hints should not be
re-evaluated after obtaining an editable instance. However, this is
only true as long as the instance obtained is for master. For branch
variants of a document we do want to re-evaluate the hints.

I have put a TODO in a separate class to make sure we don't forget to
move the key for the active project id to a shared location. We'll also
want to rename it to active_session_context_branch_id