CHANNELMGR-2014 Init RestProxyServicesManager as a Java ServiceLoader
Wicket7 deprecated the use of the wicket.properties file, which was used
to bootstrap the RestProxyServicesManager. Instead, Wicket recommends
(WICKET-5997) to use the Java ServiceLoader.

For this purpose a new Maven plugin is added to the frontend POM that
scans the class-path for classes that implement
org.apache.wicket.IInitializer and registers them in the file
META-INF/services/org.apache.wicket.IInitializer

https://issues.apache.org/jira/browse/WICKET-5997
https://docs.oracle.com/javase/9/docs/api/java/util/ServiceLoader.html