CHANNELMGR-1986 use the field name as fallback if no caption is found
The wicket editor uses the field name as fallback if no translation or caption can be found. If the caption field is there but empty, then the field label remains empty (so you can configure fields without labels). This behavior is now the same for the old and the new editor.