REPO-2094 Remove hippo:branchInfo mixin when restoring a version
The VersionRestoreToTask is used to restore a version from history. It
does this by deleting all child nodes and properties from the target
node and then selectively copying them back from the version node.
Mixins on the target node however, are not deleted. This will cause
problems if the mixin requires mandatory properties that the version
node does not have. Trying to save will then result in a constraint
violation exception.

This problem occurs when the target node has the hippo:branchInfo mixin
but the variant node not. So to fix it remove the mixin from the target.
If it is present on the version node, it will be copied back by the
DefaultCopyHandler.

I decided to only fix it for the hippo:branchInfo mixin and not
aggressively remove all mixins for two reasons:
- I might break things accidentally (e.g. in end project custom code)
- The VersionRestoreToTask is not being used anymore from workflow
  wicket plugins.