CMS-11376 Just use the node from the model
I don't understand why the workflow must be a document workflow. When I
follow the code it seems irrelevant, but instead that nullness of the
branchId is important. If it is not null we create a branch handle and
otherwise not. Since there is a lot of intricate code I'm not sure if
I miss the point somewhere. So although this change fixes the issue it
might introduce neq ones.