CMS-11400 add to-top and to-bottom controls to field controls
Orderable multiple fields now have two extra controls to move an item directly to the top or the bottom of the list. It is possible to hide the extra arrows by adding extra css classes to a field: hide-top-arrow and/o hide-bottom-arrow. This may be handy if the field is placed in a small column.