CHANNELMGR-2029 Expose function to render component for active variant
A component editor may change a component in a way that requires
re-rendering it in the site. Re-rendering a component can be achieved by
calling componentPropertiesForm.firePropertiesChanged(), but that only
works for the editor that's currently active. It should also be possible
to re-render the component from an editor that's not current active.
An editor can now do this by calling renderComponent(), which will
re-render the component with the data in the currently active tab.