CMS-11404 Make sure to get a decorated session
The session from the documentNode is not a decorated hippo session if it
is a frozen node, so it can't be cast to a HippoSession. The handle node
from the model however is always a decorated hippo node, so by using
that one to get the session we're sure it's hippo session.