CHANNELMGR-1838 Close content editor before transitioning to a new state
Fixes a race condition when opening another document that's locked. Use
case: document A was already opened in the content editor and document B
is opened, which happens to be locked. The following flow then happened:

1. uiCanExit of document A discards changes
2. uiCanExit of document B returns a resolved promise
3. document B is opened in the content editor
4. loading document B fails and sets an error in the content editor
5. uiCanExit of document A closes the content editor

Step 5 was wrong, because it reset the documentId of document B. As a
result, switch-to-content-editor did not work anymore for document B
because the documentId was undefined.

The fix is to close the content editor before returning a resolve
promise from the uiCanExit call for document A.