CHANNELMGR-1792 Highlight compound when HTML field in it is focused
The onFocus and onBlur callbacks of a CKEditor component are triggered
by a CKEditor event, which is not picked up by the change detection of
AngularJs. As a result, all state was correct but the on-focus class
in the compound's template was not added/removed yet so the color of
the compound's vertical bar did not change. Fixed by triggering a digest
cycle once the onFocus and onBlur handlers have been called.

Also removed the complex on-blur logic introduced in CHANNELMGR-1416,
which prevented the shared space toolbar from disappearing when a user
right-clicked in an HTML field. That workaround seems not necessary
anymore, probably because 12.5 uses CKEditor 4.9.