CHANNELMGR-2034 Implement and expose method to branch a document
If a document has not been branched yet, then before obtaining an
editable instance it should get branched first. I tried hard to put all
the logic in DocumentService#obtainEditableDocument(), but this is very
awkward:
After calling DocumentWorkflow#branch() the document (handle +
variants) are in a difference state and some hint values also change.
So before calling DocumentWorkflow#obtainEditableInstance we should
re-evaluate the hints and call the hintsInspector to see if we can
continue. However, as stated in the javadoc of the DocumentServiceImpl
for master documents this should not be done. This would make the code
quite incomprehensible.

A better approach is to have a REST method for every workflow action
that mutates document state. This makes error handling easier and avoids
hints going out of sync during execution of the method.

So I added a REST method that can be used to branch a document.
The client code should call it prior to requesting an editable document
for which the branch does not exist yet. By default the channel manager
code does not support branching, so the implementation of
canBranchDocument in the HintsInspector always returns NOT_BRANCHEABLE.
Similar for the BranchService implementation, branch method always
throws an UnsupportedOperationException. It's up to the module that
provides branching capabilities to implement them.