REPO-2101 Add test for asserting actions
The exising tests in the DocumentWorkflowTest are making assertions
multiple times by changing the variants under the handle. This makes it
difficult to maintain them: when the SCXML is changed you have to
keep on running the tests and fix assertion blocks one at a time.
Furthermore, testing actions that depend on nodes in version history is
not possible (or very difficult).

The new test is a parameterized test. This enables us to test the
actions for each handle state (i.e. combination of variants and their
properties). In addition, the test instantiates a SCXML workflow
executor with a mocked document handle in it. This enables us to mock
methods that normally depend on nodes in version history (e.g.
isPreviewAvailable) so that we can also test those actions.