REPO-2113 Add utility method to get a branch handle from document handle
There is a recurring use case where the document handle and the branch
handle are both needed. In those cases the same branchId is used for
constructing an instance. So it makes sense to add a utility method to
the document handle to do tha, which guarantees that the branchIds are
equal. Of course, before you can call the method the document handle
should have been initialized.