CMS-11422 Always show state icons for master branch in lists
The branch ids of unpublished and published variants under a handle are
not always the same. Example: master branch is live, while foo is being
edited. Currently the state icon in the document list is determined from
the stateSummary property of the variant node. Unfortunately that
information is not always correct if branches exist for the document.
The stateSummary is set by the DerivedDataFunction (in the repository).
If the lastModified timestamps of the unpublised and published variant
are not equal then stateSummary is set to "changed". That is only
correct if both variants are for the same branch.

So to fix it we need to get the correct published master, which could be
in version history. The DocumentHandle class can already do that and,
together with the BranchHandle class it is possible to determine the
state for master, even if the variants are in version history.

Please note that the icon reflects the state of the MASTER branch, which
doesn't have to be the same branch as the unpublished variant under the
handle. This is as per spec.