CMS-11423 Make sure the fallback session does not get logged out
during a login attempt, whether successful  or not, the fallback session
is not required to be logged out. Also instead of calling #invalidateNow
during a login attempt to force a new http session, it makes more sense
to use #replaceSession which replaces the web session without destroying
all wicket components directly. An extra advantage is that we can override
replaceSession and then suppress the logging out of the fallback session