CMS-11376 Always use the EditableWorkflow
Since #obtaineditableInstance has moved to EditableWorkflow there
is no need cast the workflow to DocumentWorkflow.

The branchModelReference and its model are always present or created
in the BranchIdModel, so the null checks can be removed.

Classes using BranchIdModel do not have to be aware of the values of
the "undefined" state only if the BranchIdModel has been defined or not.

In the BrowseService the EditorFactory is used to create a new editor.
Before that a BranchIdModel is instantiated. The test did not reflect yet
that the BranchIdModel has already been created before the editor is
started. So that has been added.