CMS-11427 BranchIdModel should always be defined
The branchIdModel should always be initialized in
org.hippoecm.frontend.plugins.cms.browse.service.BrowseService#browse.
If the editor opened by other code that code should initialize the
branchIdModel.

The branchIdModel has to be initialized before the plugin cluster
starts, because the order of creation of plugins that are configured
as sibblings cannot be determined.