CHANNELMGR-2044 order properties in groups with a unique name
Especially done for the case when there are multiple groups without label: they should all be in the same group with the new default label.

A Map is used to do this easily. The result is returned mapped in a way that it is easy to work with it in the markup.