REPO-2117 VersionHistory is missing for not yet persisted nodes
Although a node can have mix:versionable, version history can still be
missing if the node has not yet been persisted: This is because version
history is accessed by another state manager AFAIK...or it might be that
only after session#save, really the version history info gets written to
the jcr node. Either way, we thus have to anticipate that even though
there is a mix:versionable node, version history might still be missing.

Next to that I rewrote #getVariant method the part below:

if (documentVariant == null || !isBranch(documentVariant)) {
   return getFrozenVariant(variant, getDocumentVariant(UNPUBLISHED));
}
if the documentVariant is null for the variant you are looking for, I
cannot come up with any reason to search for it in version history: For
example when the live variant is missing, there is no point in looking
for it in version history AFAICS. Same goes for unpublished. And draft
is never needed to look up in version history at all. Hence I rewrote it
that when the variant is missing, we directly return empty optional