CMS-11449 Initialize branchId with master unless there is no master
Document state icons in the document list show the state of the master
documents. So when a user opens a document for editing it makes sense
to show master instead of the branch of the unpublished variant.

If a document is bootstrapped for a branch then there is no master to
show, so in that case the branchId model should updated to the branchId
of the variant.