CHANNELMGR-2062 Move datefield into new rightSidePanel.fields module
As the dateField component is now being shared between the
componentEditor and the contentEditor, it should be declared in the
context of the rightSidePanel, specifically in a new
rightSidePanel.fields module.

Apart from this I moved the componentEditorModule import from the
rightSidePanelModule to the EditComponentModule and did some
sorting on the imports and add sensible module name (also in specs).