CMS-11433 Use handle instead of frozen node
Move the initialization of the branch id model, so that the backing
object can be updated.

If a frozen node has a branch id that matches the current branch id
then the associated handle should be provided by the model service
and subsequently by the workflow and ALL related plugins.

The Documentmodel is the deepest possible level to add this business
logic. That prevents code duplication.