CHANNELMGR-2076 Move getHints to HintsUtils
I introduced lots of test failures by moving the getHints into the
EditingUtils class. These were caused by failing powermock expectations.
So instead of (yet again) changing all the expectations I moved the
method into a new util class.

In addition, I already disliked that the contextPayload trickeled into
the services, because context payload is something of the front-end. So
I also refactored the code by replacing the contextPayload by the
branchId. As a consequence, the hints no longer are augmented with the
key-value pairs of the context payload. Because the HintsInspector
implementation needs to know the branchId for checking workflow actions
the interface methods now also require a branchId, which is imo much
clearer than depending on a key-value pair of the context payload in
the hints.