CMS-11453 Show workflow menu for frozen node
In the case where the selected branch id is "master" and there is
no unpublished variant with branch master then a frozen node is shown
in the editor. Now if that frozen node does not have any branch info
the branch id will default to master and the workflow menu should be
shown instead of the versioning menu.