CHANNELMGR-2071 Replace context payload constant by service interface
The selected branchId is stored in the context payload of the
user's cms session context. How to lookup that property can be hidden by
introduction an interface. This way, modules that provide document
branching functionality can provide their custom implementation.

Also added
- javadoc explaining the intent of the interfaces.
- javadoc about the module configuration options.
Please note that this is NOT an API intended for external use.