HSTTWO-4426 Support virtual nodes (documents) for branches
Make sure that always the canonical node location is used for the
versioned node that gets decorated to a 'workspace node'. If we do not
use the canonical node location, linkrewriting for these nodes fail
because the decorated (proxied) node does return itself as canonical
node, but if it would use the virtual path location, then linkrewriting
will be done wrt the wrong path.

there is unfortunately one case where we need the virtual path: That is
in case of 'navigationStateful' links for virtual nodes. This is only
used in the context of faceted navigation which we unfortunately cannot
support with branched documents that are served from version history unless
we make many code changes which is not worth the candle (at this moment)