CMS-11475 Remove branch related mixins when copying in console
Copying node in the console also copies all the mixins. However, the
branch related mixins must not be copied because:
- It will associate the copied nodes with a branch which is probably
  not the intent of the admin using the console
- Version history is not copied. However, if a document is branched
  the version history is required to operate correctly.

I decided to implement it in the CopyDialog instead of in the
DefaultCopyHandler because DefaultCopyHandler is also used in JcrUtils
methods. So changing it's behavior could break a lot. In addition
DefaultCopyHandler is subclassed, so changing it would also have a
ripple effect that can break stuff.

I also noticed that properties of the hippo:branchInfo mixin were not
removed if I only removed the mixin. Therefore I iterate over all
properties and remove them to be sure they are all removed. I get the
properties from the node type so that we always remove all defined
mixin properties.