CHANNELMGR-2086 Do not change state of the overlayToggle in the getter
The button of the overlayToggle for the content is disabled if
isContentOverlayEnabled is false. When it is disabled then the icon
styling inside the button is overridden, so it will not render blue
even though the overlayToggle's state is true. So there is no need
to change the state value.

Same reasoning applies to the overlayToggle of the components.