CHANNELMGR-2099 Pass host group of channel as header in all REST calls
The Channel Manager Angular app fetches the JSON data about the current
channel when a channel is opened via REST. The call needs the host group
of the channel as an HTTP header so the HST knows which channel to
return. Yet that host group is part of the JSON data of a channel: a
catch 22.

Fixed via the same mechanism as the context path of a channel (which has
the same problem). All channel information is already known by the
overview, so besides the context path the overview now also passed the
host group of the opened channel to the Angular app.