CHANNELMGR-2054 Prevent editing when a component is locked or inherited
Components cannot be edited when the surrounding container is 'disabled'
(which happens when it's lock due to concurrent edits, or when the
container is inherited from another channel).

There was one snag related to naming. Angular Material includes the
following CSS rule to style disabled inputs:

  md-input-container .md-input[disabled],
  [disabled] md-input-container .md-input

The second line will render any input below an element with a 'disabled'
attribute as disabled. Consequently, the new input binding of the
componentFields and propertyGroup components cannot be called 'disabled'
since that will also render a 'disabled' attribute and hence trigger the
CSS rule even when the fields underneath are not disabled. I've
therefore used the term 'readOnly' instead.

The term 'disabled' is still used for the input bindings of the
dateField and nodeLink component since that does not clash and matches
nicely with the 'disabled' attribute for inputs. It's optional to
indicate you don't have to specify it. When ommitted it becomes
'undefined', which is falsy, so that matches nicely with inputs not
being disabled by default.